android-intermediete/SubmssionStoryApp/app/src/main/java/com/example/submssionstoryapp/view/main/MainViewModel.kt
package com.example.submssionstoryapp.view.main

import androidx.lifecycle.LiveData
import androidx.lifecycle.MutableLiveData
import androidx.lifecycle.ViewModel
import androidx.lifecycle.asLiveData
import androidx.lifecycle.viewModelScope
import androidx.paging.PagingData
import androidx.paging.cachedIn
import com.example.submssionstoryapp.data.model.ListStoryItem
import com.example.submssionstoryapp.data.model.UserModel
import com.example.submssionstoryapp.data.repository.UserRepository
import kotlinx.coroutines.launch

class MainViewModel(private val repository: UserRepository) : ViewModel() {
    private val _isLoading = MutableLiveData<Boolean>()
    val listStory: LiveData<PagingData<ListStoryItem>> =
        repository.getStory().cachedIn(viewModelScope)


    private val _message = MutableLiveData<String>()
    val message: LiveData<String> get() = _message


    fun getSession(): LiveData<UserModel> {
        return repository.getSession().asLiveData()
    }

    fun logout() {
        viewModelScope.launch {
            repository.logout()
        }
    }
}